+++ src/Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing/CommonTestInfrastructureAssemblyMarker.cs
@@ -1,0 +1,9 @@
+﻿namespace Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing
+{
+	/// <summary>
+	/// Marker class for the Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing assembly.
+	/// </summary>
+	public static class CommonTestInfrastructureMetricsCapturingAssemblyMarker
+	{
+	}
+}
