+++ src/Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing/DummyObjects/DummyAsyncCommandThatFailsMetricsCapturingStrategy.cs
@@ -1,20 +1,21 @@
 ﻿using System;
+using Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects;
 using Functional.CQS.AOP.MetricsCapturing;
 
-namespace Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects
+namespace Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing.DummyObjects
 {
 	/// <summary>
-	/// Metrics capturing strategy for <see cref="DummyCommandHandlerThatSucceeds"/>.
+	/// Metrics capturing strategy for <see cref="DummyAsyncCommandHandlerThatFails"/>.
 	/// </summary>
-	public class DummyCommandThatSucceedsMetricsCapturingStrategy : IMetricsCapturingStrategyForCommand<DummyCommandThatSucceeds, DummyCommandError>
+	public class DummyAsyncCommandThatFailsMetricsCapturingStrategy : IMetricsCapturingStrategyForCommand<DummyAsyncCommandThatFails, DummyAsyncCommandError>
 	{
 		/// <summary>
 		/// Called immediately prior to handling the command.
 		/// </summary>
 		/// <param name="parameters">The command parameters.</param>
-		public void OnInvocationStart(DummyCommandThatSucceeds parameters)
+		public void OnInvocationStart(DummyAsyncCommandThatFails parameters)
 		{
-			
+
 		}
 
 		/// <summary>
@@ -23,7 +24,7 @@ namespace Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects
 		/// <param name="parameters">The command parameters.</param>
 		/// <param name="result">The command result.</param>
 		/// <param name="timeElapsed">The time elapsed since the start of invocation.</param>
-		public void OnInvocationCompletedSuccessfully(DummyCommandThatSucceeds parameters, Result<Unit, DummyCommandError> result, TimeSpan timeElapsed)
+		public void OnInvocationCompletedSuccessfully(DummyAsyncCommandThatFails parameters, Result<Unit, DummyAsyncCommandError> result, TimeSpan timeElapsed)
 		{
 			Console.WriteLine(result.Match(_ => "succeeded", _ => "failed"));
 		}
@@ -34,9 +35,9 @@ namespace Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects
 		/// <param name="parameters">The command parameters.</param>
 		/// <param name="exception">The exception that caused the command handler to fail.</param>
 		/// <param name="timeElapsed">The time elapsed since the start of invocation.</param>
-		public void OnInvocationException(DummyCommandThatSucceeds parameters, Exception exception, TimeSpan timeElapsed)
+		public void OnInvocationException(DummyAsyncCommandThatFails parameters, Exception exception, TimeSpan timeElapsed)
 		{
-			
+
 		}
 	}
 }
