+++ src/Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing/DummyObjects/DummyQueryReturnsReferenceTypeMetricsCapturingStrategy.cs
@@ -1,20 +1,21 @@
 using System;
+using Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects;
 using Functional.CQS.AOP.MetricsCapturing;
 
-namespace Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects
+namespace Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing.DummyObjects
 {
 	/// <summary>
-	/// Metrics capturing strategy for <see cref="DummyAsyncQueryReturnsValueTypeHandler"/>.
+	/// Metrics capturing strategy for <see cref="DummyQueryReturnsReferenceType"/>.
 	/// </summary>
-	public class DummyAsyncQueryReturnsValueTypeMetricsCapturingStrategy : IMetricsCapturingStrategyForQuery<DummyAsyncQueryReturnsValueType, DummyAsyncQueryReturnsValueTypeResult>
+	public class DummyQueryReturnsReferenceTypeMetricsCapturingStrategy : IMetricsCapturingStrategyForQuery<DummyQueryReturnsReferenceType, DummyQueryReturnsReferenceTypeResult>
 	{
 		/// <summary>
 		/// Called immediately prior to handling the query.
 		/// </summary>
 		/// <param name="parameters">The query parameters.</param>
-		public void OnInvocationStart(DummyAsyncQueryReturnsValueType parameters)
+		public void OnInvocationStart(DummyQueryReturnsReferenceType parameters)
 		{
-
+			
 		}
 
 		/// <summary>
@@ -23,9 +24,9 @@ namespace Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects
 		/// <param name="parameters">The query parameters.</param>
 		/// <param name="result">The query result.</param>
 		/// <param name="timeElapsed">The time elapsed since the start of invocation.</param>
-		public void OnInvocationCompletedSuccessfully(DummyAsyncQueryReturnsValueType parameters, DummyAsyncQueryReturnsValueTypeResult result, TimeSpan timeElapsed)
+		public void OnInvocationCompletedSuccessfully(DummyQueryReturnsReferenceType parameters, DummyQueryReturnsReferenceTypeResult result, TimeSpan timeElapsed)
 		{
-
+			
 		}
 
 		/// <summary>
@@ -34,7 +35,7 @@ namespace Functional.CQS.AOP.CommonTestInfrastructure.DummyObjects
 		/// <param name="parameters">The query parameters.</param>
 		/// <param name="exception">The exception that caused the query handler to fail.</param>
 		/// <param name="timeElapsed"></param>
-		public void OnInvocationException(DummyAsyncQueryReturnsValueType parameters, Exception exception, TimeSpan timeElapsed)
+		public void OnInvocationException(DummyQueryReturnsReferenceType parameters, Exception exception, TimeSpan timeElapsed)
 		{
 
 		}
