+++ src/Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing/Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Functional.CQS.AOP.CommonTestInfrastructure\Functional.CQS.AOP.CommonTestInfrastructure.csproj" />
+    <ProjectReference Include="..\Functional.CQS.AOP.IoC.PureDI.MetricsCapturing\Functional.CQS.AOP.IoC.PureDI.MetricsCapturing.csproj" />
+    <ProjectReference Include="..\Functional.CQS.AOP.MetricsCapturing\Functional.CQS.AOP.MetricsCapturing.csproj" />
+  </ItemGroup>
+
+</Project>
