+++ src/Functional.CQS.sln
@@ -25,6 +25,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functional.CQS.AOP.IoC.Simp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functional.CQS.AOP.IoC.SimpleInjector.MetricsCapturing.Tests", "Functional.CQS.AOP.IoC.SimpleInjector.MetricsCapturing.Tests\Functional.CQS.AOP.IoC.SimpleInjector.MetricsCapturing.Tests.csproj", "{3B104AC4-23F1-4944-8B69-A13A812A3730}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing", "Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing\Functional.CQS.AOP.CommonTestInfrastructure.MetricsCapturing.csproj", "{8246DD38-6775-46D5-AF05-768647908E60}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -67,6 +69,10 @@ Global
 		{3B104AC4-23F1-4944-8B69-A13A812A3730}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B104AC4-23F1-4944-8B69-A13A812A3730}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3B104AC4-23F1-4944-8B69-A13A812A3730}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8246DD38-6775-46D5-AF05-768647908E60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8246DD38-6775-46D5-AF05-768647908E60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8246DD38-6775-46D5-AF05-768647908E60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8246DD38-6775-46D5-AF05-768647908E60}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
